Change from 32-bit to 64-bit integer vertices

+++ app/src/main/java/navigation/Edge.java
@@ -4,7 +4,7 @@ import collections.enumflags.EnumFlags;
 
 import java.io.Serializable;
 
-record Edge(int from, int to, float distance, float maxSpeed, EnumFlags<EdgeRole> roles) implements Serializable {
+record Edge(long from, long to, float distance, float maxSpeed, EnumFlags<EdgeRole> roles) implements Serializable {
     public boolean hasRole(EdgeRole role) {
         return roles.isSet(role);
     }
